eSTREAM fix: block size can be modified (still has to be fixed in estream_cipher.h)
eSTREAM was proposed for ciphers with block length 128, but some has 256 bits block implementation as well. We have block length in the configuration file, but the block length was fixed in the source code, so the users input was ignored. The problem is not fixed for IV and key, that should have different lengths as well.

The old implementation of TEA (should be deprecated, will be removed during this set of fixes) has smaller block size. Therefore it was encrypting zero blocks. This behaviour changed from EAC 3 to EAC 4.